Rename tutorial component as MutationTutorial.
